requirements.txt
numpy==1.18.5
tensorflow-cpu==2.3.0
streamlit==0.70.0
